Bug #6 from Jan 2016 Addendum: Increase FTP Pacer max to 200%
This may or may not be the complete fix. The UI has been updated to
allow for up to 200% FTP, and it seems to have an affect in 3D Cycling.

However, during testing there were some very noticeable issues with
pacer performance (both a 50% FTP pacer and 200% FTP pacer were riding
at the same speed). Due to other bug reports, I believe this is a
pre-existing condition, however it prevents proper testing &
verification of this particular bug, so I'm going to investigate a bit
more.