Bug #5 on Jan 2016 Addendum: Update AeT -> AnT.
Throughout the code, UI, and documentation there are references to "AT"
being Aerobic Threshold, but it was actually intended to be Anaerobic
Threshold. This change replaces all references to "AeT" and "Aerobic
Threshold" with "AnT" and "Anaerobic Threshold"; this includes variable
names and XML. Care was taken to ensure that upon reading an "AeT"
ridersDB that it will be loaded properly and converted to "AnT" when
it's saved out again.