Update Wifi Configuration Dialog to use latest console exe.
The new combination of RM1 + console.exe + racermate.dll + 2820 firmware
seems to work great for a computer that primarily uses wifi to connect to
their network!

One caveat: After successfully configuring one or more wifi controllers,
RM1 needs to be restarted in order for the new controllers to be detected.