Hardware Dialog now automatically updates when it sees a new or lost trainer.
When the dialog is notified of a trainer being found or lost, it automatically calls 'StartQuickScan()' to re-init the trainers and update the UI.

Sometimes this might happen a little too quickly, and the FirmWareVersion
of the trainer will still be set to 0. There's a request sent to Larry to
see if that bug can be resolved in the racermate library.