RM1: Fix updating Velotron gears based on rider's profile.
The previous code would only update the physical Velotron's gearing if
there was a new rider on that Velotron. This change now also checks to see
if the rider's chainrings and cogs match what is already set on the
Velotron, and will update the gearing if needed.

Also added some mutex protection around the OnTrainerLost and
OnTrainerFound, as I ran into a couple cases where the callbacks were
removed while we were in the midst of trying to call them (so the
callbacks when from being non-null to being null, and causing an
exception).