RM1: Fix loading rider database with 11 gears.
The crash we were seeing was actually in our error message box because it
wasn't able to setup a keyboard binding. This change disables the key
binding since I wasn't sure what it was supposed to do.

The rider database was failing to load because we didn't make an array big
enough to store all 11 gears. This array is now tied to the
RM1.MAX_REAR_GEARS constant that is set to 11. This new flexbility should
allow us to increase all the way up to 20 if needed (that limit is imposed
by other parts of the software and the RacerMate library).

Also made sure the CrankSet would allow for the same flexibility as well
if for some reason we decide to increase that from 3 gears up to 5.

Bumped the version number to 4.1.0.6
* Hopefully just the AssemblyInfo needed to be updated!